Rename actions-types to actionTypes
