+++ src/redux/user/reducer.js
@@ -1,4 +1,4 @@
-import { ACTION_TYPES } from './actions-types';
+import { ACTION_TYPES } from './actionTypes';
 
 export const defaultUserState = {
     isOrgAdmin: undefined
