+++ src/utilities/store.js
@@ -6,7 +6,7 @@ import promise from 'redux-promise-middleware';
 import SourcesReducer, { defaultSourcesState } from '../redux/sources/reducer';
 import UserReducer, { defaultUserState } from '../redux/user/reducer';
 import { updateQuery } from './urlQuery';
-import { ACTION_TYPES } from '../redux/sources/actions-types';
+import { ACTION_TYPES } from '../redux/sources/actionTypes';
 
 export const urlQueryMiddleware = store => next => action => {
     if (action.type === ACTION_TYPES.LOAD_ENTITIES_PENDING) {
